Include find_hip in utils. Also change import statements in _init_
Change location for find_hip.
Delete read_HIP.py

Change import statements to reflect this.
