+++ barycorrpy/__init__.py
@@ -1,9 +1,8 @@
-from . import utc_tdb
-from . import PINT_erfautils 
-from . import read_HIP
-from . import utils
-from . import sample_script
+from .utc_tdb import *
+from .PINT_erfautils import *
+from .utils import *
+from .sample_script import *
 from .barycorrpy import get_BC_vel,BCPy,exposure_meter_BC_vel
 
 
-__version__ = '0.2'
+__version__ = '0.2devel'
